Move OfflineSigner interface to signer module
